requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.31.0
SQLAlchemy>=2.0.36
pydantic>=2.11.0
PyJWT>=2.9.0
passlib>=1.7.4
Jinja2>=3.1.5
python-dotenv>=1.0.1

google-genai>=0.6.0

scikit-learn>=1.5.0
joblib>=1.4.2
pandas>=2.2.2

slowapi>=0.1.9